[Bug fix] Average income-outcome visual / Float Enterpretation
Also: comment formatting

+++ core/src/com/ttocsneb/income/save/Transaction.java
@@ -9,10 +9,10 @@ public class Transaction {
 	/**
 	 * The amount of money received.
 	 */
-	public long income;
+	public float income;
 	
 	/**
 	 * The amount of money lost.
 	 */
-	public long expense;
+	public float expense;
 }
